requirements.txt
Babel==2.7.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.3
django-address==0.2.1
django-allauth==0.39.1
django-ckeditor==5.7.1
django-countries==5.5
django-js-asset==1.2.2
django-phonenumber-field==3.0.1
django-phonenumbers==1.0.1
djangorestframework==3.10.2
idna==2.8
oauthlib==3.0.2
phonenumbers==8.10.18
Pillow==6.1.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
